refactor: change name to reflect only string use
