+++ src/helpers/stringHelpers.ts
