+++ src/pages/showcase/lists/PaginationBasicExample.tsx
@@ -11,7 +11,7 @@ import {
 import { generateFakerData } from '@/helpers/fakerHelpers';
 import BrothImage from '@/components/image/BrothImage';
 import { BrothTypography } from '@/components/typography/BrothTypography';
-import { limitLength } from '@/helpers/textHelpers';
+import { limitLength } from '@/helpers/stringHelpers';
 import { ChangeEvent, useMemo, useState } from 'react';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
